@@ -1,8 +1,11 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <UserSecretsId>aspnet-IdentityServer-5FE82F6F-0884-4871-829E-BCFE700497A2</UserSecretsId>
+    <OutputType>Exe</OutputType>
+    <SignAssembly>false</SignAssembly>
+    <StartupObject>IdentityServer.Program</StartupObject>
   </PropertyGroup>
 
   <ItemGroup>
@@ -10,12 +13,22 @@
   </ItemGroup>
 
 
-  <ItemGroup>
-    <PackageReference Include="IdentityModel" Version="3.6.1" />
-    <PackageReference Include="IdentityServer4.AspNetIdentity" Version="2.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.7" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="2.0.2" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.0.3" PrivateAssets="All" />
+  <ItemGroup> 
+    <PackageReference Include="IdentityModel" Version="4.4.0" /> 
+    <PackageReference Include="IdentityServer4" Version="4.1.0" /> 
+    <PackageReference Include="IdentityServer4.AspNetIdentity" Version="4.1.0" /> 
+    <PackageReference Include="IdentityServer4.EntityFramework" Version="4.1.0" /> 
+    <PackageReference Include="log4net" Version="2.0.11" /> 
+    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.1.8" /> 
+    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.8" /> 
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.8" /> 
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1.8">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference> 
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="3.1.8" /> 
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.8" /> 
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.8" /> 
     <PackageReference Include="Serilog" Version="2.6.0" />
     <PackageReference Include="Serilog.Extensions.Logging" Version="2.0.2" />
     <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
@@ -26,6 +39,10 @@
     <DotNetCliToolReference Include="Microsoft.EntityFrameworkCore.Tools.DotNet" Version="2.0.2" />
     <DotNetCliToolReference Include="Microsoft.Extensions.SecretManager.Tools" Version="2.0.1" />
     <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.3" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\WebAppSettings\VikingWebAppSettings.csproj" />
   </ItemGroup>
 
 </Project>
