@@ -59,15 +59,24 @@ namespace IdentityServer.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [Authorize(Roles = "Access Manager")]
-        public async Task<IActionResult> Create([Bind("Id,Name")] Group organization)
+        public async Task<IActionResult> Create([Bind("Name,ShortName")] Group group)
         {
             if (ModelState.IsValid)
             {
-                _context.Add(organization);
+                _context.Add(group);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            return View(organization);
+
+            return View(group);
+        }
+
+        public IActionResult VerifyUniqueGroupName(string groupName)
+        {
+            if (_context.Group.Any(g => g.Name == groupName))
+                return Json($"A group named {groupName} already exists");
+
+            return Json(true); 
         }
 
         // GET: Organizations/Edit/5
@@ -78,26 +87,32 @@ namespace IdentityServer.Controllers
                 return NotFound();
             }
 
-            var organizationEditDetails = await _context.Group.Include("GroupAssignments").Select(org => new GroupDetailsViewModel
-                {
-                    Name = org.Name,
-                    Id = org.Id,
-                    UserList = _context.Users.Select(u => new UserSelectedViewModel
-                    {
-                        Id = u.Id,
-                        Name = u.UserName,
-                        Selected = org.GroupAssignments.Any(oa => oa.UserId == u.Id)
-                    }).ToList()
-                })
-                .SingleOrDefaultAsync(m => m.Id == id);
-             
-            if (organizationEditDetails == null)
+            var groupEditDetials = await _context.Group.Include("GroupAssignments").Select(org => new GroupDetailsViewModel
             {
-                return NotFound();
-            }
-
-            return View(organizationEditDetails);
+                Name = org.Name,
+                Id = org.Id,
+                UserList = _context.Users.Select(u => new UserSelectedViewModel
+                {
+                    Id = u.Id,
+                    Name = u.UserName,
+                    Selected = org.GroupAssignments.Any(oa => oa.UserId == u.Id)
+                }).ToList(),
+                Children = org.Children.Select(g => new GroupDetailsViewModel
+                {
+                    Name = g.Name,
+                    Id = g.Id
+                }
+                ).ToList()
+            })
+            .SingleOrDefaultAsync(m => m.Id == id);
+             
+        if (groupEditDetials == null)
+        {
+            return NotFound();
         }
+
+        return View(groupEditDetials);
+    }
 
         // POST: Organizations/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
@@ -105,16 +120,16 @@ namespace IdentityServer.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [Authorize(Roles = "Access Manager")]
-        public async Task<IActionResult> Edit(long id, [Bind("Id,Name,ShortName")] GroupDetailsViewModel organizationDetails, [Bind] IEnumerable<UserSelectedViewModel> usersSelected)
+        public async Task<IActionResult> Edit(long id, [Bind("Id,Name,Description")] GroupDetailsViewModel groupDetails, [Bind] IEnumerable<UserSelectedViewModel> usersSelected)
         {
-            if (id != organizationDetails.Id)
+            if (id != groupDetails.Id)
             {
                 return NotFound();
             }
 
-            var organization = await _context.Group.Include("GroupAssignments").SingleOrDefaultAsync(m => m.Id == id);
+            var group = await _context.Group.Include("GroupAssignments").SingleOrDefaultAsync(m => m.Id == id);
 
-            if (organization == null)
+            if (group == null)
             {
                 return NotFound();
             }
@@ -124,10 +139,10 @@ namespace IdentityServer.Controllers
             {
                 try
                 {
-                    organization.Name = organizationDetails.Name;
-                    organization.ShortName = organizationDetails.ShortName;
+                    group.Name = groupDetails.Name;
+                    group.Description = groupDetails.Description;
 
-                    organization.UpdateUserOrganizations(usersSelected);
+                    group.UpdateUserOrganizations(usersSelected);
                     /*
                     foreach(UserSelectedViewModel user in usersSelected)
                     {
@@ -152,12 +167,12 @@ namespace IdentityServer.Controllers
                         }
                     }*/
                     
-                    _context.Update(organization);
+                    _context.Update(group);
                     await _context.SaveChangesAsync(); 
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!OrganizationExists(organization.Id))
+                    if (!OrganizationExists(group.Id))
                     {
                         return NotFound();
                     }
@@ -170,7 +185,7 @@ namespace IdentityServer.Controllers
                 return RedirectToAction(nameof(Index));
             }
 
-            return View(organizationDetails);
+            return View(groupDetails);
         }
 
         // GET: Organizations/Delete/5
