@@ -25,10 +25,13 @@ namespace IdentityServer.Controllers
     public class UserRolesController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly ILogger _logger;
 
-        public UserRolesController(ApplicationDbContext context)
+        public UserRolesController(ApplicationDbContext context,
+            ILogger<AccountController> logger)
         {
             _context = context;
+            _logger = logger;
         }
          
         // GET: UserRoles 
@@ -101,7 +104,7 @@ namespace IdentityServer.Controllers
         // POST: UserRoles/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Access Manager")]
+        [Authorize(Roles = Config.AdminRoleName)]
         public ActionResult Edit(UserRolesViewModel id, IFormCollection collection)
         {
             try
@@ -121,6 +124,16 @@ namespace IdentityServer.Controllers
                     }
                     else if(!check && listUserRoles.Any(ur => ur.RoleId == userRole.Id))
                     {
+                        //Safety check, make sure we do not remove the last admin user from the admin role
+                        if(userRole.Name == Config.AdminRoleName)
+                        {
+                            bool otherAdminUsers = _context.UserRoles.Where(ur => ur.RoleId == Config.AdminRoleId && ur.UserId != User.Id).Any();
+                            if(otherAdminUsers== false)
+                            {
+                                _logger.LogWarning("Cannot remove the last admin user");
+                                continue; 
+                            }
+                        }
                         var urToRemove = listUserRoles.First(ur => ur.RoleId == userRole.Id && ur.UserId == User.Id);
                         _context.UserRoles.Remove(urToRemove);
                     }
