@@ -19,15 +19,17 @@ namespace IdentityServer.Models.UserViewModels
         [MaxLength(450)]
         public string Name { get; set; }
 
-        [Required(AllowEmptyStrings = false)]
-        [Display(Name = "Short Unique Identifier", Description = "A unique short identifier for the organization.  This value is used for affiliation claims placed in tokens.")]
-        [MaxLength(64)]
-        public string ShortName { get; set; }
+        [Display(Name = "Description", Description = "Information about the group")]
+        [MaxLength(2048)]
+        public string Description { get; set; }
 
         /// <summary>
         /// All users in the database, the selected property is true if they are a member of the organization
         /// </summary>
         [Display(Name = "Users")]
         public List<UserSelectedViewModel> UserList { get; set; }
+
+        [Display(Name = "Children")]
+        public List<GroupDetailsViewModel> Children { get; set; }
     }
 }
