@@ -1,7 +1,6 @@
 ﻿@using System.Collections.Generic
 @using System.Linq
-@using Microsoft.AspNetCore.Http
-@using Microsoft.AspNetCore.Http.Authentication
+@using Microsoft.AspNetCore.Http 
 @model LoginViewModel
 @inject SignInManager<ApplicationUser> SignInManager
 
