@@ -40,21 +40,24 @@
                     </label>
                 </div>
             </div>
+            <!-- No need to expose auto generated values
+
+             <div class="form-group">
+                 <label asp-for="PasswordHash" class="control-label"></label>
+                 <input asp-for="PasswordHash" class="form-control" />
+                 <span asp-validation-for="PasswordHash" class="text-danger"></span>
+             </div>
+             <div class="form-group">
+                 <label asp-for="SecurityStamp" class="control-label"></label>
+                 <input asp-for="SecurityStamp" class="form-control" />
+                 <span asp-validation-for="SecurityStamp" class="text-danger"></span>
+             </div>
             <div class="form-group">
-                <label asp-for="PasswordHash" class="control-label"></label>
-                <input asp-for="PasswordHash" class="form-control" />
-                <span asp-validation-for="PasswordHash" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="SecurityStamp" class="control-label"></label>
-                <input asp-for="SecurityStamp" class="form-control" />
-                <span asp-validation-for="SecurityStamp" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="ConcurrencyStamp" class="control-label"></label>
-                <input asp-for="ConcurrencyStamp" class="form-control" />
-                <span asp-validation-for="ConcurrencyStamp" class="text-danger"></span>
-            </div>
+                 <label asp-for="ConcurrencyStamp" class="control-label"></label>
+                 <input asp-for="ConcurrencyStamp" class="form-control" />
+                 <span asp-validation-for="ConcurrencyStamp" class="text-danger"></span>
+             </div>
+             -->
             <div class="form-group">
                 <label asp-for="PhoneNumber" class="control-label"></label>
                 <input asp-for="PhoneNumber" class="form-control" />
