@@ -1,11 +1,11 @@
 ﻿@model IEnumerable<IdentityServer.Models.ApplicationUser>
 
 @{
-    ViewData["Title"] = "Index";
+    ViewData["Title"] = "Users";
     Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
-<h2>Index</h2>
+<h2>Users</h2>
 
 <p>
     <a asp-action="Create">Create New</a>
