@@ -5,7 +5,7 @@
 <h3>@ViewData["Message"]</h3>
 
 <address>
-    Marc Lab<br />
+    Jones Lab<br />
     Moran Eye Center <br />
     65 Mario Cappechii Drive
     Salt lake City, UT 84132<br />
